requirements.txt
Flask==2.3.3
selenium==4.15.2
sympy==1.12
numpy==1.24.3
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
flask-socketio==5.3.6
gevent==23.9.1
webdriver-manager==4.0.1